backend/requirements.txt
# --- Web Server ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# --- Async HTTP & Database ---
httpx>=0.27.0
aiosqlite>=0.20.0

# --- Core Logic & Networking ---
websockets==11.0.3
requests==2.32.3
python-dotenv==1.0.1

# --- Blockchain Interaction ---
solana==0.35.0
solders==0.21.0

# --- Communication APIs ---
python-telegram-bot==21.10
tweepy==4.15.0
google-generativeai>=0.7.0
